refactor: 직관적인 파일명으로 수정
