+++ src/app/SDKProvider.tsx
@@ -1,15 +1,15 @@
-'use client';
+"use client";
 
-import { useSession } from 'next-auth/react';
-import { useSpotifySDK } from '@/hooks/useSpotifySDK';
+import { useSession } from "next-auth/react";
+import { useSpotifySDK } from "@/hooks/useInitSpotifyPlayer";
 
 export default function SDKProvider() {
-    const { data: session } = useSession();
+  const { data: session } = useSession();
 
-    const accessToken =
-        session && typeof session === 'object' ? session.accessToken : null;
+  const accessToken =
+    session && typeof session === "object" ? session.accessToken : null;
 
-    useSpotifySDK(accessToken);
+  useSpotifySDK(accessToken);
 
-    return null;
+  return null;
 }
